iter5: error handling
