+++ cmd/server/main.go
@@ -18,7 +18,7 @@ func main() {
 func run() error {
 	config, err := NewConfig()
 	if err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	storage := repository.NewMemory()
