--- internal/agent/repository/storage.go
@@ -1,11 +1,0 @@
-package repository
-
-import "runtime"
-
-type Storage interface {
-	Save(memStats *runtime.MemStats) error
-	IncrementCounter()
-	ClearCounter()
-	GetGauges() map[string]gauge
-	GetCounters() map[string]counter
-}
